requirements.txt
pytesseract : extracts text from images
pillow : helps open/resize images
tesseract not jst python wrapper
torch
torchvision
transformers
sentence-transformers
faiss-cpu